requirements.txt
# PUU Universal Hub Dependencies
fastapi==0.110.0
uvicorn==0.27.1
psycopg[binary]==3.1.18
jinja2==3.1.3
python-dotenv==1.0.1
httpx==0.27.0
mcp==0.1.0
typing-extensions==4.9.0
python-multipart==0.0.9
